Refactor SyncStorage trait.
So that we can reference account logs when the sync feature is not
enabled using the StorageEventLogs trait.
